Rewrite disk cache implementation to cache only on download.
Fixes: #7458, #7573, #6265, #6630, #7938 and #6934

This will allow cache to consistently work for
server and gateways. Range GET requests will
be cached in the background after the request
is served from the backend.

- All cached content is automatically bitrot protected.

- Avoid ETag verification if cache-control header
is set and cached content is still valid.

- This PR changes the cache backend format, and all existing
content will be migrated to the new format. Until the data is
migrated completely, all content will be served from the backend.